remove broken reset feature, bring queries in their own file
