+++ hooks/useGraphData.ts
@@ -13,7 +13,6 @@ import { ParseNode } from "@/utils/ggraph";
 
 export function useGraphData() {
   const searchParams = useSearchParams();
-  const router = useRouter();
 
   const [graphData, setGraphData] = useState<GraphDataWithMetadata>({
     nodes: [],
@@ -83,15 +82,12 @@ export function useGraphData() {
     }
   };
 
-  const resetGraph = (nodeId: string) => {
-    router.push(`/?path=${nodeId}`);
-  };
-
   // fetch and update data based on query parameters
   const fetchDataFromQueryParams = async () => {
     try {
       const myQuery = searchParams.get("path");
-      if (myQuery !== null) {
+
+      if (myQuery) {
         const nodeIds = myQuery.split(",");
         const parsedNodes = await fetchAndParseNodes(nodeIds);
 
@@ -124,6 +120,5 @@ export function useGraphData() {
     breadcrumbs,
     addBreadcrumb,
     removeBreadcrumbsFromIndex,
-    resetGraph,
   };
 }
